recognition/vietocr/requirements.txt
torch==2.0.0
torchvision==0.15.1
opencv-python
vietocr
pyyaml
Pillow